Rename mappings.js to api.js
